# Optimize node encoding/decoding with stream-vbyte format