@tac0turtle 2023-02-08
Would love to open an issue for this. It looks awesome, maybe someone from outside can pick it up